requirements.txt
pytesseract
pdf2image
Pillow
pdfplumber
unidecode
langchain
sentence-transformers
chromadb
numpy